Add files via upload
Write a program to find out area or volume of an shape/object by using one function name as
FUN-AREA only. If the function takes one argument then it will calculate area of a circle, two
argument for area of rectangle and three argument volume of a box. If no argument is given to
the function, then it will display appropriate message..